Move internal files in order to keep the path structure used by the other apps - src/ -> src/app
