+++ apps/network/src/app/network/network_manager.py
