+++ apps/network/src/app/routes/network.py
