+++ apps/network/src/app/routes/nodes.py
