+++ apps/network/src/app/workers/worker.py
